fix(pareto_persistence): rm rPref param, write & test base & rPref functions
NOTE: Instead of using the additional .use_rPref parameter, this commit splits
the pareto_persistence function into a base calculator and an rPref calculator,
with the original calling one or the other according as rPref is installed.
This restores the pre-parameter behavior elsewhere and allows both calculator
functions to be tested (provided rPref is installed).